+++ src/System.Threading.AccessControl/tests/AclTests.cs
@@ -1,0 +1,10 @@
+﻿namespace System.Threading.Tests
+{
+    public class AclTests
+    {
+        protected string GetRandomName()
+        {
+            return Guid.NewGuid().ToString("N");
+        }
+    }
+}
