+++ src/System.Threading.AccessControl/tests/System.Threading.AccessControl.Tests.csproj
@@ -5,6 +5,8 @@
   <ItemGroup>
     <Compile Include="$(CommonPath)\CoreLib\Interop\Windows\Kernel32\Interop.Constants.cs" Link="Common\CoreLib\Interop\Windows\Kernel32\Interop.Constants.cs" />
     <Compile Include="$(CommonPath)\CoreLib\Interop\Windows\Kernel32\Interop.MAX_PATH.cs" Link="Common\CoreLib\Interop\Windows\Kernel32\Interop.MAX_PATH.cs" />
+    <Compile Include="AclTests.cs" />
+    <Compile Include="MutexAclTests.cs" />
     <Compile Include="MutexSecurityTests.cs" />
     <Compile Include="ThreadingAclExtensionsTests.cs" />
   </ItemGroup>
